feat: Add metrics response schema and event handling for automation mappings
- Introduced MetricsResponse schema for system metrics including event activity, mapping statistics, lamp states, and system connectivity.
- Enhanced MappingRepository with methods for creating, updating, and deleting mappings.
- Implemented Events controller for managing event injection and mapping operations.
- Developed EventCache service to prevent duplicate event processing.
- Created EventService for processing button events and executing corresponding actions.
- Added LegacyAdapter for converting legacy MQTT message formats to modern schemas.
- Configured Docker Compose for production environment with MySQL and MQTT settings.

+++ backend/src/models/light.py
@@ -1,6 +1,6 @@
 """Light model (luzes table - legacy/compatibility)."""
 from datetime import datetime
-from sqlalchemy import Boolean, Column, DateTime, Integer, String,BigInteger
+from sqlalchemy import Boolean, Column, DateTime, Integer, String,BigInteger, BIGINT
 from sqlalchemy.orm import relationship
 
 from src.infra.db import Base
@@ -11,7 +11,7 @@ class Light(Base):
     
     __tablename__ = "luzes"
     
-    id = Column( BigInteger(unsigned=True), primary_key=True, index=True, autoincrement=True)
+    id = Column( BIGINT(unsigned=True), primary_key=True, index=True, autoincrement=True)
     lampada = Column(String(50), unique=True, nullable=False, index=True)
     estado = Column(Boolean, default=False, nullable=False)
     data_de_atualizacao = Column(
